Split the integration test `test_dictionaries` into 4 separate tests.
