--- dbms/tests/integration/test_dictionaries/configs/dictionaries/.gitignore
@@ -1,4 +1,0 @@
-*
-!.gitignore
-!source.tsv
-!dictionary_preset*
