+++ dbms/tests/integration/test_dictionaries_all_layouts_and_sources/configs/config.xml
