+++ dbms/tests/integration/test_dictionaries_complex_key_cache_string/configs/config.xml
