+++ dbms/tests/integration/test_dictionaries_depend_on_dictionaries/configs/users.xml
