+++ dbms/tests/integration/test_dictionaries_null_value/configs/dictionaries/cache.xml
@@ -1,0 +1,113 @@
+    <yandex>
+       <dictionary>
+            <name>cache</name>
+
+            <source>
+                <clickhouse>
+                    <host>localhost</host>
+                    <port>9000</port>
+                    <user>default</user>
+                    <password></password>
+                    <db>test</db>
+                    <table>source</table>
+                </clickhouse>
+            </source>
+
+            <lifetime>0</lifetime>
+
+            <layout>
+                <cache><size_in_cells>128</size_in_cells></cache>
+            </layout>
+
+            <structure>
+                <id>
+                    <name>id</name>
+                </id>
+                
+                <attribute>
+                    <name>UInt8_</name>
+                    <type>UInt8</type>
+                    <null_value>1</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>UInt16_</name>
+                    <type>UInt16</type>
+                    <null_value>1</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>UInt32_</name>
+                    <type>UInt32</type>
+                    <null_value>1</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>UInt64_</name>
+                    <type>UInt64</type>
+                    <null_value></null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>Int8_</name>
+                    <type>Int8</type>
+                    <null_value>-1</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>Int16_</name>
+                    <type>Int16</type>
+                    <null_value>-1</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>Int32_</name>
+                    <type>Int32</type>
+                    <null_value>-1</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>Int64_</name>
+                    <type>Int64</type>
+                    <null_value>-1</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>Float32_</name>
+                    <type>Float32</type>
+                    <null_value>2.71828</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>Float64_</name>
+                    <type>Float64</type>
+                    <null_value>2.71828</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>String_</name>
+                    <type>String</type>
+                    <null_value>implicit-default</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>Date_</name>
+                    <type>Date</type>
+                    <null_value>2015-11-25</null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>DateTime_</name>
+                    <type>DateTime</type>
+                    <null_value></null_value>
+                </attribute>
+                
+                <attribute>
+                    <name>Parent</name>
+                    <type>UInt64</type>
+                    <hierarchical>true</hierarchical>
+                    <null_value>0</null_value>
+                </attribute>
+            </structure>
+        </dictionary>
+    </yandex>
