+++ dbms/tests/integration/test_dictionaries_select_all/configs/dictionaries/.gitignore
@@ -1,0 +1,3 @@
+*
+!.gitignore
+!source.tsv
