+++ dbms/tests/integration/test_dictionaries_update_and_reload/configs/dictionaries/cache_xypairs.xml
